Added selecting inventory slots and using items
I shifted the berries up and left by 5 in preparation for making the borders on the slot selected thicker like in minecraft but I couldn't figure out the border so I went with the dot. I left it in case someone wants to figure it out. For now you can change it back to centered on line 333 by getting rid of the -5 in the first 2 args.